+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Add support for registering, updating, and listing smart contracts that are
+deployed external to CDP.
+
 - Add `network_id` to `WalletData` so that it is saved with the seed data and surfaced via the export function
 
 ### [0.12.1] - 2024-12-10