+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Add ability to import external wallets into CDP via a BIP-39 mnemonic phrase, as a 1-of-1 wallet
 - Add ability to import WalletData files exported by the NodeJS CDP SDK
 - Deprecate `Wallet.load_seed` method in favor of `Wallet.load_seed_from_file`
-- Deprecate `Wallet.saveSeed()` method in favor of `Wallet.save_seed_to_file`
+- Deprecate `Wallet.save_seed` method in favor of `Wallet.save_seed_to_file`
 
 ### [0.12.1] - 2024-12-10